requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
unittest-parametrize==1.0.0
